+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * Added landmark class for signal-related queries.
   * Added support to parse OpenDRIVE signals.
   * Added junction class as queryable object from waypoint
+  * Added timeout to World Tick
   * Added simple physical map generation from standalone OpenDRIVE data
   * Added support for generating walker navigation on server-side
   * Added support for new geometry: `spiral`, `poly3`, and `paramPoly3`